Hack in a task preview by task id
This doesn't render meta data since it's not clear which task
collection to use, as well as which position the task is in it. While
the latter could be found out, its useless anyway.

Also renders no buttons. But thats only a hack to avoid navigating away
from the preview. There are still useless calls that get a task
collection object (in this case one with the id=1 --> therefore a TC
with this id has to exist!).
This is just to appease the template which does a reverse lookup of an
URL, which needs the tc.id and tc_task_id…)

A new template is needed. Or this preview should only be embedded into
the admin interface anyway.