Basic session support
Disallows skipping ahead and backtracking during a quiz.
Stores progress in the session data.
At then end a hashvalue gets calculated that includes the log and the
current time. This is presented to the student.
It is also stored in the DB, together with the timestamp and the log in
plain text.